requirements.txt
adal==1.2.2
azure-batch==8.0.0
azure-common==1.1.25
azure-mgmt-batch==7.0.0
azure-mgmt-resource==8.0.1
azure-mgmt-storage==8.0.0
azure-nspkg==3.0.2
azure-storage==0.36.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
isodate==0.6.0
msrest==0.6.11
msrestazure==0.6.2
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
